Configurable Rounded Corners: Introduced a new user-configurable setting for window corner radius, which dynamically applies rounded corners to the title bar background, button backgrounds, and the overall window outline. Corners are automatically squared off when the window is maximized.
Standardized Icon Painting: Refactored the drawing logic for all button icons to use a standardized approach, scaling them into a fixed 18x18 unit canvas. This ensures consistent icon appearance across different button types and sizes, simplifying future icon design and maintenance.
Enhanced Geometric Precision: Migrated numerous geometry and size calculations from integer (int) to floating-point (qreal) types. This change improves the precision of UI element positioning and rendering, reducing potential visual artifacts and enabling smoother animations.
Button Corner Awareness: Added isLeftmost and isRightmost flags to individual buttons, allowing for more granular control over how corners are rendered for buttons positioned at the edges of a button group. This enables more sophisticated visual styling for button clusters.

+++ src/AppMenuButtonGroup.h
@@ -79,7 +79,7 @@ public:
     qreal opacity() const;
     void setOpacity(qreal value);
 
-    int visibleWidth() const;
+    qreal visibleWidth() const;
 
     bool isMenuOpen() const;
     bool menuLoadedOnce() const;
@@ -158,7 +158,7 @@ private:
     bool m_animationEnabled;
     QVariantAnimation *m_animation;
     qreal m_opacity;
-    int m_visibleWidth;
+    qreal m_visibleWidth;
     QPointer<QMenu> m_currentMenu;
     int m_buttonIndexWaitingForPopup = -1;
     int m_buttonIndexOfMenuToCache = -1;
